src/requirements.txt
asgiref==3.5.2
Django==4.1.2
django-environ==0.9.0
django-filter==22.1
django-multiselectfield==0.1.12
django-phonenumber-field==7.0.0
djangorestframework==3.14.0
gunicorn==20.1.0
importlib-metadata==5.0.0
Markdown==3.4.1
phonenumbers==8.12.57
Pillow==9.2.0
psycopg2==2.9.4
pytz==2022.5
sqlparse==0.4.3
tzdata==2022.5
waitress==2.1.2
zipp==3.9.0
